ci: Add publishing GitHub Actions workflow (#20)
* Add a GitHub Actions workflow that will build the package, verify the
package with twine, list the contents of the sdist and wheel for manual
inspection, and then either upload the the distributions to TestPyPI
or PyPI. The workflow runs on push events to main or for tags, on
pull request events to main (for testing), and on demand with workflow
dispatch.

Publishing is controlled through different events. Publishing to
TestPyPI if there is a git tag pushed or if there is a workflow dispatch
in which publishing is selected. Publishing to PyPI is only through
GitHub releases.